modified:   huffman_encoder.c Joining code in functions 	modified:   huffman_functions.c 	modified:   huffman_functions.h

+++ huffman_functions.c
@@ -14,7 +14,7 @@
 
 #include "huffman_functions.h"
 
-unsigned long pow_int(int base, int exp)
+unsigned long pow_int(unsigned long base, unsigned long exp)
 {
 	int res = 1;
 	for (; exp > 0; exp--)
